# Branch-less process for binary arithmetic functions like intDiv and mod