--- a/src/Functions/FunctionBinaryArithmetic.h
+++ b/src/Functions/FunctionBinaryArithmetic.h
@@ -269,10 +269,17 @@ struct BinaryOperation
             if (right_nullmap)
             {
                 for (size_t i = 0; i < size; ++i)
-                    if ((*right_nullmap)[i])
-                        c[i] = ResultType();
-                    else
-                        apply<op_case>(a, b, c, i);
+                {
+                    if constexpr (std::is_integral_v<ResultType>)
+                        c[i] = (!!(*right_nullmap)[i]) * ResultType() + (!(*right_nullmap)[i]) * apply<op_case>(a, b, c, i);
+                     else
+                     {
+                        if ((*right_nullmap)[i])
+                            c[i] = ResultType();
+                        else
+                            apply<op_case>(a, b, c, i);
+                     }
+                }
             }
             else
                 for (size_t i = 0; i < size; ++i)
@@ -284,12 +291,14 @@ struct BinaryOperation
 
 private:
     template <OpCase op_case>
-    static void apply(const A * __restrict a, const B * __restrict b, ResultType * __restrict c, size_t i)
+    static ResultType apply(const A * __restrict a, const B * __restrict b, ResultType * __restrict c, size_t i)
     {
         if constexpr (op_case == OpCase::Vector)
             c[i] = Op::template apply<ResultType>(a[i], b[i]);
         else
             c[i] = Op::template apply<ResultType>(*a, b[i]);
+
+        return c[i];
     }
 };