--- a/src/Functions/intDiv.cpp
+++ b/src/Functions/intDiv.cpp
@@ -39,11 +39,12 @@ struct DivideIntegralByConstantImpl
         {
             if (right_nullmap)
             {
+                //std::cerr << "right_nullmap" << std::endl;
+                /// For better performance, try to use branch free code for numeric types(i.e. cond ? a : b --> !!cond * a + !cond * b)
                 for (size_t i = 0; i < size; ++i)
-                    if ((*right_nullmap)[i])
-                        c[i] = ResultType();
-                    else
-                        apply<op_case>(a, b, c, i);
+                {
+                    c[i] = (!!(*right_nullmap)[i]) * ResultType() + (!(*right_nullmap)[i]) * apply<op_case>(a, b, c, i);
+                }
             }
             else
                 for (size_t i = 0; i < size; ++i)
@@ -80,12 +81,14 @@ struct DivideIntegralByConstantImpl
 
 private:
     template <OpCase op_case>
-    static void apply(const A * __restrict a, const B * __restrict b, ResultType * __restrict c, size_t i)
+    static ResultType apply(const A * __restrict a, const B * __restrict b, ResultType * __restrict c, size_t i)
     {
         if constexpr (op_case == OpCase::Vector)
             c[i] = Op::template apply<ResultType>(a[i], b[i]);
         else
             c[i] = Op::template apply<ResultType>(*a, b[i]);
+
+        return c[i];
     }
 };