--- a/tests/performance/modulo.xml
+++ b/tests/performance/modulo.xml
@@ -2,4 +2,5 @@
     <query>SELECT number % 128 FROM numbers(300000000) FORMAT Null</query>
     <query>SELECT number % 255 FROM numbers(300000000) FORMAT Null</query>
     <query>SELECT number % 256 FROM numbers(300000000) FORMAT Null</query>
+    <query>SELECT number % toNullable(number+1) FROM numbers(300000000) FORMAT Null</query>
 </test>