@zhanglistar 2024-09-12
For some design, this way is not OK. For example, 1/0 will throw execption breaks the rule, we should make 1/0 = NULL. I will try to get this.